fixed pin issue with leading 0

+++ app/src/main/java/com/example/gihealth/ui/onboarding/CreatePinScreen.kt
@@ -132,7 +132,7 @@ fun CreatePinScreen(navController: NavController, onPinCreated: () -> Unit) {
                         else -> {
                             //No error message and goes to next part of app
                             errorMessage = ""
-                            userInfoViewModel.saveUserPin(pin.toInt())
+                            userInfoViewModel.saveUserPin(pin)
                             navController.navigate("enter_pin")
                             onPinCreated()
                         }
